+++ app/src/test/kotlin/io/mszymanski/orknux/server/action/ActionNodeRunnerTest.kt
@@ -230,7 +230,7 @@ class ActionNodeRunnerTest(
             """
             mutation {
               createWorkspaceConnection(input: {
-                workspaceId: $workspaceId, name: "Slack", type: SLACK_SOCKET_MODE, url: "https://slack.com/api"
+                workspaceId: $workspaceId, name: "Slack", type: SLACK
               }) { id }
             }
             """,